Refactor rotate and have only one command with a --ccw flag for counter-clockwise rotation. By default, rotate is clockwise
