+++ crates/nu-command/src/default_context.rs
@@ -85,7 +85,6 @@ pub fn create_default_context(cwd: impl AsRef<Path>) -> EngineState {
             Reject,
             Reverse,
             Rotate,
-            RotateCounterClockwise,
             Select,
             Shuffle,
             Skip,
