--- crates/nu-command/src/filters/rotate/mod.rs
@@ -1,5 +1,0 @@
-mod command;
-mod counter_clockwise;
-
-pub use command::Rotate;
-pub use counter_clockwise::SubCommand as RotateCounterClockwise;
